+++ src/main/java/one/talon/model/ApplicationAPIKey.java
@@ -156,23 +156,23 @@ public class ApplicationAPIKey {
 
   public static final String SERIALIZED_NAME_TIME_OFFSET = "timeOffset";
   @SerializedName(SERIALIZED_NAME_TIME_OFFSET)
-  private Integer timeOffset;
+  private Long timeOffset;
 
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED_BY = "createdBy";
   @SerializedName(SERIALIZED_NAME_CREATED_BY)
-  private Integer createdBy;
+  private Long createdBy;
 
   public static final String SERIALIZED_NAME_ACCOUNT_I_D = "accountID";
   @SerializedName(SERIALIZED_NAME_ACCOUNT_I_D)
-  private Integer accountID;
+  private Long accountID;
 
   public static final String SERIALIZED_NAME_APPLICATION_I_D = "applicationID";
   @SerializedName(SERIALIZED_NAME_APPLICATION_I_D)
-  private Integer applicationID;
+  private Long applicationID;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -269,7 +269,7 @@ public class ApplicationAPIKey {
   }
 
 
-  public ApplicationAPIKey timeOffset(Integer timeOffset) {
+  public ApplicationAPIKey timeOffset(Long timeOffset) {
     
     this.timeOffset = timeOffset;
     return this;
@@ -282,17 +282,17 @@ public class ApplicationAPIKey {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "100000", value = "A time offset in nanoseconds associated with the API key. When making a request using the API key, rule evaluation is based on a date that is calculated by adding the offset to the current date. ")
 
-  public Integer getTimeOffset() {
+  public Long getTimeOffset() {
     return timeOffset;
   }
 
 
-  public void setTimeOffset(Integer timeOffset) {
+  public void setTimeOffset(Long timeOffset) {
     this.timeOffset = timeOffset;
   }
 
 
-  public ApplicationAPIKey id(Integer id) {
+  public ApplicationAPIKey id(Long id) {
     
     this.id = id;
     return this;
@@ -304,17 +304,17 @@ public class ApplicationAPIKey {
   **/
   @ApiModelProperty(example = "34", required = true, value = "ID of the API Key.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
 
-  public ApplicationAPIKey createdBy(Integer createdBy) {
+  public ApplicationAPIKey createdBy(Long createdBy) {
     
     this.createdBy = createdBy;
     return this;
@@ -326,17 +326,17 @@ public class ApplicationAPIKey {
   **/
   @ApiModelProperty(example = "280", required = true, value = "ID of user who created.")
 
-  public Integer getCreatedBy() {
+  public Long getCreatedBy() {
     return createdBy;
   }
 
 
-  public void setCreatedBy(Integer createdBy) {
+  public void setCreatedBy(Long createdBy) {
     this.createdBy = createdBy;
   }
 
 
-  public ApplicationAPIKey accountID(Integer accountID) {
+  public ApplicationAPIKey accountID(Long accountID) {
     
     this.accountID = accountID;
     return this;
@@ -348,17 +348,17 @@ public class ApplicationAPIKey {
   **/
   @ApiModelProperty(example = "13", required = true, value = "ID of account the key is used for.")
 
-  public Integer getAccountID() {
+  public Long getAccountID() {
     return accountID;
   }
 
 
-  public void setAccountID(Integer accountID) {
+  public void setAccountID(Long accountID) {
     this.accountID = accountID;
   }
 
 
-  public ApplicationAPIKey applicationID(Integer applicationID) {
+  public ApplicationAPIKey applicationID(Long applicationID) {
     
     this.applicationID = applicationID;
     return this;
@@ -370,12 +370,12 @@ public class ApplicationAPIKey {
   **/
   @ApiModelProperty(example = "54", required = true, value = "ID of application the key is used for.")
 
-  public Integer getApplicationID() {
+  public Long getApplicationID() {
     return applicationID;
   }
 
 
-  public void setApplicationID(Integer applicationID) {
+  public void setApplicationID(Long applicationID) {
     this.applicationID = applicationID;
   }
